refactor: 💡 final refactor
✅ Closes: #10
